CDLmotif

Documentation for CDLmotif.

CDLmotif.activation_record

CDLmotif.activations_topology

CDLmotif.data_2_dummy

CDLmotif.pvalue2score

CDLmotif.sample_m

CDLmotif.activation_record — Type

A record data structure for each dna sequence in the dataset T1: whatever Int type T2: smaller Int type to save space; e.g. Int16, Int32

CDLmotif.activations_topology — Type

struct for inferring the topology of the activations construction: K: total number of filters (from all of the trainings) maxfillen: maximum length of the filters CminusFsminus1: cover_length minus the smallest filter length minus 1 which is C-|fₛ|-1; this is the maximum length that the filters can be apart from each other in the cover area

Note: 1. I am using "topology" in the loose sense here 2. T should be Int16 or Int32 since - K: total number of filters is typically less than 100 - Use Int16 if overlap/nonoverlap/nextto count is less than 32,768

CDLmotif.data_2_dummy — Method

get the set of dummy-vectors from a set of dna-strings the dummy-vectors are all of same length (for now)

CDLmotif.pvalue2score — Method

pval2score(pwm, pval, ϵ=1e-1, k=10, bg=[.25,.25,.25,.25])

Returns the highest score(M,pval) of a pwm such that p-value is greater or equal to pval.

Input:

pwm: a 4 x m matrix

pval: a p-value; e.g. pval = 1e-3

ϵ: initial granularity (optional)

k: Refinement parameter (optional)

bg: multinomial background (optional)

Output

alpha: the highest score-threshold

CDLmotif.sample_m — Method

Sample the ith mode of the motif, a mixturepfam_motifs